requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.7
defusedxml==0.7.1
dill==0.3.3
distlib==0.3.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
furl==2.1.2
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.1
mistune==0.8.4
msal==1.11.0
msrest==0.6.21
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
num2words==0.5.10
numpy==1.20.3
oauthlib==3.1.1
openpyxl==3.0.7
orderedmultidict==1.0.1
packaging==20.9
pandas==1.2.4
pandoc==1.0.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
ply==3.11
portalocker==1.7.1
progressbar==2.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xgboost==1.4.1
yarl==1.6.3